requirements.txt
captum==0.6.0
imageio==2.31.1
ipython==8.24.0
jupyter-core==5.7.2
krippendorff==0.6.1
matplotlib==3.7.2
numpy==1.25.1
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
scikit-image==0.21.0
scikit_learn==1.3.0
scipy==1.13.1
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
typing_extensions==4.12.0
